Hotfixes

+++ public/assets/js/common-bundle.js
@@ -403,7 +403,7 @@ toastr.options = {
   window.calcularSubtotalItem = function(){
 
     var precio_unitario = parseFloat( $('#precio_unitario').val() );
-    var cantidad = parseInt( $('#cantidad').val() );
+    var cantidad = parseFloat( $('#cantidad').val() );
     var porc_iva = parseFloat( $('#porc_iva').val() );
     var monto_iva = parseFloat( $('#item_iva_amount').val() );
     
@@ -444,7 +444,7 @@ toastr.options = {
   window.calcularConIvaManual = function(){
     
     var precio_unitario = parseFloat( $('#precio_unitario').val() );
-    var cantidad = parseInt( $('#cantidad').val() );
+    var cantidad = parseFloat( $('#cantidad').val() );
     var monto_iva = parseFloat( $('#item_iva_amount').val() );
     
     if( !monto_iva ) {
